Move urlFilter function into settings.
